Moves request_handler code to .cc implementation
